renames diana api
