+++ diana/diana-key-value/src/main/java/org/jnosql/diana/kv/DefaultKeyValueEntityProvider.java
@@ -14,7 +14,7 @@
  *   Otavio Santana
  *
  */
-package org.jnosql.diana.key;
+package org.jnosql.diana.kv;
 
 import jakarta.nosql.Value;
 import jakarta.nosql.kv.KeyValueEntity;
