+++ diana/diana-key-value/src/main/java/org/jnosql/diana/kv/query/DefaultKeyValueQueryParser.java
@@ -14,7 +14,7 @@
  *   Otavio Santana
  *
  */
-package org.jnosql.diana.key.query;
+package org.jnosql.diana.kv.query;
 
 import jakarta.nosql.QueryException;
 import jakarta.nosql.Value;
