--- diana/diana-key-value/src/main/resources/META-INF/services/jakarta.nosql.key.KeyValueEntity$KeyValueEntityProvider
@@ -1,1 +1,0 @@
-org.jnosql.diana.key.DefaultKeyValueEntityProvider
