+++ diana/diana-key-value/src/main/resources/META-INF/services/jakarta.nosql.kv.KeyValueQueryParser
@@ -1,0 +1,1 @@
+org.jnosql.diana.kv.query.DefaultKeyValueQueryParser
