+++ diana/diana-key-value/src/test/java/org/jnosql/diana/kv/query/DefaultKeyValueQueryParserTest.java
@@ -14,7 +14,7 @@
  *   Otavio Santana
  *
  */
-package org.jnosql.diana.key.query;
+package org.jnosql.diana.kv.query;
 
 import jakarta.nosql.kv.BucketManager;
 import jakarta.nosql.kv.KeyValueEntity;
