+++ diana/diana-key-value/src/test/java/org/jnosql/diana/kv/query/GetQueryParserTest.java
@@ -14,7 +14,7 @@
  *   Otavio Santana
  *
  */
-package org.jnosql.diana.key.query;
+package org.jnosql.diana.kv.query;
 
 import jakarta.nosql.QueryException;
 import jakarta.nosql.kv.BucketManager;
